+++ HISTORY.md
@@ -1,7 +1,36 @@
 Release History
 ===============
 
-0.0.1 (2023-88888888888888888888888-88888888888888888888888)
+TODOS/WISHES
+------------
+1. add SERIAL execution as method wait_all_piped! paired up with wait_all_parallel() 
+2. add meta cumulative funks
+
+0.0.7 (2024-01-10)
 -------------------
-**NEW**
-- 
+- add terminate_all
+
+0.0.6 (2024-01-10)
+-------------------
+- BIG REF! for thread_instance
+- use QThread (terminate+SecondaryStart)
+
+0.0.5 (2023-12-19)
+-------------------
+- fix nothread
+
+0.0.4 (2023-11-01)
+-------------------
+- add PARAM__NOTHREAD
+
+0.0.3 (2023-11-01)
+-------------------
+- add check_results_all
+
+0.0.2 (2023-10-13)
+-------------------
+- apply new singleton version
+
+0.0.1 (2023-10-03)
+-------------------
+- first working release!